+++ src/enums/pageEnum.ts
@@ -6,9 +6,10 @@ export enum PageEnum {
   REDIRECT = '/redirect',
   REDIRECT_NAME = 'Redirect',
   // 首页
-  BASE_HOME = '/dashboard',
+  BASE_HOME = '/homeStatic/index',
   //首页跳转默认路由
-  BASE_HOME_REDIRECT = '/dashboard/console',
+  BASE_HOME_REDIRECT = '/homeStatic/index',
   // 错误
   ERROR_PAGE_NAME = 'ErrorPage',
 }
+