+++ src/layout/components/Menu/index.vue
@@ -123,6 +123,7 @@
         if (/http(s)?:/.test(key)) {
           window.open(key);
         } else {
+          console.log(key);
           router.push({ name: key });
         }
       }