+++ src/views/homeStatic/Admin.vue
@@ -1,0 +1,11 @@
+<template>
+  <div>
+    超管首页
+  </div>
+</template>
+
+<script lang="ts" setup>
+</script>
+
+<style lang="less" scoped>
+</style>