+++ src/views/homeStatic/Requestor.vue
@@ -1,0 +1,11 @@
+<template>
+  <div>
+    需求方首页
+  </div>
+</template>
+
+<script lang="ts" setup>
+</script>
+
+<style lang="less" scoped>
+</style>