+++ src/views/homeStatic/Supplier.vue
@@ -1,0 +1,12 @@
+<template>
+  <div>
+    供应商首页
+  </div>
+</template>
+
+<script lang="ts" setup>
+
+</script>
+
+<style lang="less" scoped>
+</style>